Reorganize service objects under G5AuthenticatableApi::Services
