+++ spec/lib/g5_authenticatable_api/helpers/rails_spec.rb
@@ -20,7 +20,7 @@ describe G5AuthenticatableApi::Helpers::Rails, type: :controller do
                                access_token: token_value)
     end
     before do
-      allow(G5AuthenticatableApi::TokenValidator).to receive(:new).
+      allow(G5AuthenticatableApi::Services::TokenValidator).to receive(:new).
         and_return(token_validator)
     end
 
@@ -30,7 +30,7 @@ describe G5AuthenticatableApi::Helpers::Rails, type: :controller do
 
       it 'initializes the token validator correctly' do
         authenticate_api_user!
-        expect(G5AuthenticatableApi::TokenValidator).to have_received(:new).
+        expect(G5AuthenticatableApi::Services::TokenValidator).to have_received(:new).
           with(request.params,
                an_instance_of(ActionDispatch::Http::Headers),
                request.env['warden'])
